requirements-vercel.txt
# Vercel-optimized requirements
Flask==2.3.3
Flask-CORS==4.0.0
numpy==1.24.3
scikit-learn==1.3.0
joblib==1.3.2
python-dotenv==1.0.0
requests==2.31.0
PyJWT==2.8.0
psycopg2-binary==2.9.7
